requirements.txt
datasets
numpy
pandas
psutil
python-dotenv
requests
scikit-learn
torch
transformers